start of ES2015 application structure
