+++ frontend/imports/startup/client/index.js
@@ -1,0 +1,2 @@
+// Import to load these templates
+import '../../ui/client/widgets/ethtokens.js';
