+++ frontend/package.json
@@ -5,5 +5,8 @@
     "test": "meteor test --port 3100 --once --driver-package dispatch:mocha-phantomjs",
     "test:watch": "TEST_WATCH=1 meteor test --port 3100 --driver-package dispatch:mocha-phantomjs",
     "start": "meteor run"
+  },
+  "dependencies": {
+    "meteor-node-stubs": "~0.2.0"
   }
 }
